hamed/add_mt5_settings_unit_test
- add mt5 settings unit test

+++ lib/api/models/mt5_settings_model.dart
@@ -46,7 +46,7 @@ abstract class MT5SettingsModel extends BaseModel {
   final String group;
 
   /// Client leverage (from 1 to 1000).
-  final double leverage;
+  final int leverage;
 
   /// Login ID of the user's MT5 account.
   final String login;
